LayoutTests/platform/glib/media/media-source/worker/media-managedmse-worker-expected.txt


EXPECTED (ManagedMediaSource.canConstructInDedicatedWorker == 'true'), OBSERVED 'undefined' FAIL
RUN(video.disableRemotePlayback = true)
error message from worker: MSE not supported FAIL
END OF TEST
